+++ route53/metadata.rb
@@ -4,6 +4,4 @@ maintainer_email "support@hw-ops.com"
 license          "Apache 2.0"
 description      "Installs/Configures route53"
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          "0.4.0"
-
-depends 'xml'
+version          "0.5.6"