+++ route53/test/integration/cookbooks/route53_test/metadata.rb
@@ -4,6 +4,6 @@ maintainer_email 'aaron@hw-ops.com'
 license          'Apache 2.0'
 description      'Installs/Configures route53_wrapper test cookbook'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.1.0'
+version          '0.2.0'
 
 depends 'route53'