+++ route53/test/integration/cookbooks/route53_test/recipes/default.rb
@@ -37,32 +37,25 @@ route53_record node[:records][:generic_record][:name] do
   value                 node[:records][:generic_record][:value]
   type                  node[:records][:generic_record][:type]
   ttl                   node[:records][:generic_record][:ttl]
+  region                node[:route53][:region]
   zone_id               node[:route53][:zone_id]
-  aws_access_key_id     node[:route53][:aws_access_key_id]
-  aws_secret_access_key node[:route53][:aws_secret_access_key]
-  overwrite             true
   action                :create
-  mock                  true
 end
 
 route53_record node[:records][:alias_record][:name] do
-  alias_target          node[:records][:alias_record][:alias_target]
   type                  node[:records][:alias_record][:type]
   zone_id               node[:route53][:zone_id]
-  aws_access_key_id     node[:route53][:aws_access_key_id]
-  aws_secret_access_key node[:route53][:aws_secret_access_key]
-  overwrite             true
+  ttl                   node[:records][:generic_record][:ttl]
+  region                node[:route53][:region]
   action                :create
   only_if               { node[:records][:alias_record][:run] }
-  mock                  true
 end
 
 route53_record "#{node[:records][:generic_record][:name]}_delete" do
   name                  node[:records][:generic_record][:name]
   type                  node[:records][:generic_record][:type]
   zone_id               node[:route53][:zone_id]
-  aws_access_key_id     node[:route53][:aws_access_key_id]
-  aws_secret_access_key node[:route53][:aws_secret_access_key]
+  ttl                   node[:records][:generic_record][:ttl]
+  region                node[:route53][:region]
   action                :delete
-  mock                  true
 end